# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfoSec Harness",
    "version": "2.0.0.dev0"
  },
  "paths": {
    "/api/run-page": {
      "get": {
        "summary": "Run Page",
        "operationId": "run_page_api_run_page_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch Id"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "total_tokens",
                    "input_tokens",
                    "output_tokens",
                    "cost_usd",
                    "wall_time_s"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric"
            }
          },
          {
            "name": "lower",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lower"
            }
          },
          {
            "name": "upper",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Upper"
            }
          },
          {
            "name": "upper_inclusive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Upper Inclusive"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/experiments/{experiment_id}": {
      "get": {
        "summary": "Experiment Detail",
        "operationId": "experiment_detail_api_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "population",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "operational",
                    "demo",
                    "legacy"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Population"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExperimentCase": {
        "properties": {
          "case_name": {
            "type": "string",
            "title": "Case Name"
          },
          "repetition": {
            "type": "integer",
            "title": "Repetition"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "scores": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Scores"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "latency_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency S"
          }
        },
        "type": "object",
        "required": [
          "case_name",
          "repetition",
          "passed",
          "scores",
          "cost_usd",
          "latency_s"
        ],
        "title": "ExperimentCase"
      },
      "ExperimentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Metrics"
          },
          "cases": {
            "items": {
              "$ref": "#/components/schemas/ExperimentCase"
            },
            "type": "array",
            "title": "Cases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent",
          "metrics",
          "cases"
        ],
        "title": "ExperimentDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit",
          "as_of"
        ],
        "title": "RunPage"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "complete",
          "needs_info",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "description": "Persisted lifecycle of one finding's triage run."
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "cwe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwe"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "inconclusive_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inconclusive Reason"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "environment_scope": {
            "type": "string",
            "title": "Environment Scope"
          },
          "early_exit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Early Exit"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens"
          },
          "latency_s": {
            "type": "number",
            "title": "Latency S"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "telemetry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunTelemetry"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "fingerprint",
          "title",
          "repo_url",
          "revision",
          "cwe",
          "severity",
          "status",
          "verdict",
          "confidence",
          "inconclusive_reason",
          "priority",
          "priority_score",
          "environment_scope",
          "early_exit",
          "cost_usd",
          "total_tokens",
          "cache_read_tokens",
          "latency_s",
          "created_at",
          "telemetry",
          "phase"
        ],
        "title": "RunSummary"
      },
      "RunTelemetry": {
        "properties": {
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "wall_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Time S"
          },
          "agent_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Time S"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "known_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Cost Usd"
          },
          "accounting_complete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accounting Complete"
          },
          "cost_accounting_complete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Accounting Complete"
          },
          "known_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Tokens"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "cost_coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Coverage"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RunTelemetry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}